Add services folder with map service, add map service spec, remove refactor commented code
